quest-36 Add new quest reward endpoints (#39)
* quest-36 enriching quests with rewards add new streaming endpoint

* fixing tests

* fixing tests

* update streaming endpoints to return quest with reward

+++ src/main/scala/models/rewards/CreateRewards.scala
@@ -8,7 +8,7 @@ import java.time.LocalDateTime
 
 case class CreateReward(
   questId: String,
-  rewardValue: BigDecimal
+  rewardValue: Int
 )
 
 object CreateReward {
